迁移ViewModel，集成串流状态
